Add Heraldicon blazonry integration
The !hd command allows parsing and rendering blazonry via
heraldicon.org, display the resulting PNG, and provide a link to where
the parsed arms can be edited directly.

In the case of errors the confusing string is indicated and
suggestions are displayed.

The !hd-options command displays all options that can be put as
keywords in front of the blazon with the !hd command.

Details: https://github.com/mirnovov/heraldtron/issues/1